C# "Can't find specified the file"
This just a simple project I'm making to rename files dynamically when
they're put in the drag drop box. Meaning we don't have to use an
absolute path. we do it all in the Drag Drop box that appears on the
form.

It can't find the file. Does anyone know what the problem is please. I'm
new to this.